requirements.txt
Django==1.7
Jinja2==2.7.3
MarkupSafe==0.23
argparse==1.2.1
django-extensions==1.4.0
django-jinja==1.0.4
django-pipeline==1.3.25
djangorestframework==2.4.2
flake8==2.2.3
futures==2.1.6
mccabe==0.2.1
mistune==0.4
nodeenv==0.11.1
pep8==1.5.7
pyflakes==0.8.1
six==1.7.3
wsgiref==0.1.2
